Stop formatting citekeys as Org-ref links in ROAM_KEY
The presence of the `cite:` in the file-metadata tag `#+ROAM_KEY:` seems to
have been warranted at some point in Org-roam, but this doesn’t seem to be the
case anymore.  Since the Org-ref link doesn’t work in that context, there’s
even less of a reason to keep it around.