orb-ignore-bibtex-store-link-functions
Org Ref defines the function `org-ref-bibtex-store-link' to store links
to a BibTeX buffer, e.g. with `org-store-link'.  At the same time, Org
ref requires `ol-bibtex' library, which defines `org-bibtex-store-link'
to do the same.  When creating a note with `orb-edit-notes' from a
BibTeX buffer, for example by calling `org-ref-open-bibtex-notes', the
initiated `org-capture' process implicitly calls `org-store-link'.  The
latter loops through all the functions for storing links, and if more
than one function can store links to the location, the BibTeX buffer in
this particular case, the user will be prompted to choose one.  This is
definitely annoying, hence ORB will advise all functions in
`orb-ignore-bibtex-store-link-functions` to return nil to trick
`org-capture' and get rid of the prompt.