(fix) Add citekey to ROAM_REFS in new note
Provide `orb--add-ref` hook function for
`org-roam-capture-new-node-hook` to add the citation key to ROAM_REFS in
a newly created note.

The behaviour of `org-roam-node-at-point` has changed in 409a147
breaking the previous way citation keys were added to ROAM_REFS, see #193.

Should the upstream changes be reverted, this may be reverted as well.