Ensure bibtex-completion is initialized
`bibtex-completion-format-entry` fails if
`bibtex-completion-display-formats-internal` is `nil`, which is the case if
bibtex completion has not been initialized via `(bibtex-completion-init)`.

This happened to me because I install `bibtex-completion` without `helm-bibtex`
or `ivy-bibtex`, calling commands of either of those packages seems to
initialize it. When it is not initialized, that results in the error message

    s-format: Wrong type argument: arrayp,
    nil

because bibtex tries to format a `nil` string.

Calling `(bibtex-completion-init)` seems to be cheap after subsequent
invocations, but I'm not sure that calling it in every `orb-format-entry` is the
best solution or we can somehow make sure to call it only once.