Initialize bibtex with together with org-roam-bibtex
Bibtex needs to be initialized via `(bibtex-completion-init)` at least once to
set the value of `bibtex-completion-display-formats-internal`, which is
necessary for the formatting via `bibtex-completion-format-entry` to work.

Calling that function at every invocation of `orb-format-entry` is probably
unnecessary overhead, so we I it from there to the mode initialization. It is
still called at every invocation of `orb-insert-generic`, but that is consistent
with what the `helm-bibtex` and `ivy-bibtex` commands do and probably for
checking that the bibliography still exists (but I'm not sure).